refactor(gateway): extract skill-hint helpers from run.py (slice 17 of #54962)
